pyproject.toml
[tool.poetry]
name = "ai_agent"
version = "0.0.1"
description = ""
authors = ["Orbis <infectores@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain-community = "^0.3.20"
langchain-gigachat = "^0.3.9"
langgraph = "^0.3.24"
langgraph-checkpoint = "^2.0.24"
langgraph-cli = "^0.1.83"
langgraph-sdk = "^0.1.61"
rapidfuzz = "^3.12.2"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"
black = "^25.1.0"
ruff = "^0.11.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "ai_agent.main:main"
chat = "ai_agent.main:chat"